+++ src/classical_register.jl
@@ -52,12 +52,6 @@ function apply!(r::Register, indices::Base.AbstractVecOrTuple{Int}, operation; k
     r
 end
 
-@deprecate apply!(r::Register, operation, indices::Base.AbstractVecOrTuple{Int}; kwargs...) apply!(r, indices, operation; kwargs...)
-
-function apply!(r::Register, indices::Base.AbstractVecOrTuple{Int}, operation::Base.AbstractVecOrTuple{Int}; kwargs...)
-    error("`apply!(::Register, ...)` requires two more arguments in specific order -- `indices` (a vector or tuple of qubit indices) and `operation` (the quantum operation acting on the qubits)")
-end
-
 function apply!(r::Register, m::sMX)
     _, res = projectXrand!(r,m.qubit)
     m.bit!=0 && (bitview(r)[m.bit] = !iszero(res))